requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
cryptography==43.0.1
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==5.1.2
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
    #   social-auth-app-django
django-allauth==65.0.2
    # via -r requirements.in
django-cors-headers==4.5.0
    # via -r requirements.in
django-debug-toolbar==4.4.6
    # via -r requirements.in
django-filter==24.3
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via
    #   -r requirements.in
    #   djoser
djoser==2.2.3
    # via -r requirements.in
drf-spectacular==0.27.2
    # via -r requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
ruff==0.6.9
    # via -r requirements.in
social-auth-app-django==5.4.2
    # via djoser
social-auth-core==4.5.4
    # via social-auth-app-django
sqlparse==0.5.1
    # via
    #   django
    #   django-debug-toolbar
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via requests
